shifted tests for ingest command into it's own file. it's about to grow
